# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: Kratos-CMS 后台管理API
    description: Kratos-CMS 后台管理API
    contact:
        name: tx7do
        url: https://github.com/tx7do/kratos-cms
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/kratos-cms/blob/master/LICENSE
    version: "1.0"
paths:
    /admin/v1/attachments:
        get:
            tags:
                - AttachmentService
            description: 获取附件列表
            operationId: AttachmentService_ListAttachment
            parameters:
                - name: page
                  in: query
                  description: 当前页码
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  description: 每页的行数
                  schema:
                    type: integer
                    format: int32
                - name: query
                  in: query
                  description: 与过滤参数
                  schema:
                    type: string
                - name: or
                  in: query
                  description: 或过滤参数
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: array
                    items:
                        type: string
                - name: nopaging
                  in: query
                  description: 是否不分页
                  schema:
                    type: boolean
                - name: field_mask
                  in: query
                  description: 字段掩码
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListAttachmentResponse'
        post:
            tags:
                - AttachmentService
            description: 创建附件
            operationId: AttachmentService_CreateAttachment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateAttachmentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Attachment'
    /admin/v1/attachments/{id}:
        get:
            tags:
                - AttachmentService
            description: 获取附件数据
            operationId: AttachmentService_GetAttachment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Attachment'
        put:
            tags:
                - AttachmentService
            description: 更新附件
            operationId: AttachmentService_UpdateAttachment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Attachment'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Attachment'
        delete:
            tags:
                - AttachmentService
            description: 删除附件
            operationId: AttachmentService_DeleteAttachment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        Attachment:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                name:
                    type: string
                path:
                    type: string
                fileKey:
                    type: string
                thumbPath:
                    type: string
                mediaType:
                    type: string
                suffix:
                    type: string
                width:
                    type: integer
                    format: int32
                height:
                    type: integer
                    format: int32
                size:
                    type: string
                type:
                    type: integer
                    format: int32
                createTime:
                    type: string
                updateTime:
                    type: string
                deleteTime:
                    type: string
            description: 附件
...
        CreateAttachmentRequest:
            type: object
            properties:
                attachment:
                    $ref: '#/components/schemas/Attachment'
                operatorId:
                    type: integer
                    format: uint32
            description: 创建附件 - 请求
...
        ListAttachmentResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Attachment'
                total:
                    type: integer
                    format: int32
            description: 附件列表 - 回应
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh_token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...